scaffold hello view into component
