+++ app/(tabs)/hello.tsx
@@ -1,21 +1,5 @@
-import { StyleSheet, View } from 'react-native';
-import { ThemedText } from '@/components/ThemedText';
+import HelloWorld from '@/components/HelloWorld';
 
 export default function HelloScreen() {
-  return (
-    <View style={styles.container}>
-      <ThemedText style={styles.text}>Hello World</ThemedText>
-    </View>
-  );
+  return <HelloWorld />;
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  text: {
-    fontSize: 24,
-  },
-});
